+++ cmd/main.go
@@ -267,6 +267,22 @@ func main() {
 		os.Exit(1)
 	}
 
+	if err := (&controller.MaintenanceCalendarReconciler{
+		Client:  mgr.GetClient(),
+		Clients: clientPool,
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "MaintenanceCalendar")
+		os.Exit(1)
+	}
+
+	if err := (&controller.MaintenanceScheduleReconciler{
+		Client:  mgr.GetClient(),
+		Clients: clientPool,
+	}).SetupWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "MaintenanceSchedule")
+		os.Exit(1)
+	}
+
 	if err := (&webhook.OrganizationValidator{}).SetupWebhookWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create webhook", "webhook", "OrganizationValidator")
 		os.Exit(1)
@@ -334,6 +350,16 @@ func main() {
 
 	if err := (&webhook.CustomInfoKeyValidator{}).SetupWebhookWithManager(mgr); err != nil {
 		setupLog.Error(err, "unable to create webhook", "webhook", "CustomInfoKeyValidator")
+		os.Exit(1)
+	}
+
+	if err := (&webhook.MaintenanceCalendarValidator{}).SetupWebhookWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create webhook", "webhook", "MaintenanceCalendarValidator")
+		os.Exit(1)
+	}
+
+	if err := (&webhook.MaintenanceScheduleValidator{}).SetupWebhookWithManager(mgr); err != nil {
+		setupLog.Error(err, "unable to create webhook", "webhook", "MaintenanceScheduleValidator")
 		os.Exit(1)
 	}